% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatStructured.R
\name{formatStructured}
\alias{formatStructured}
\title{formatStructured: Function to add structured data into the workflow.}
\usage{
formatStructured(dataOCC, type, varsOld, varsNew, projection, boundary)
}
\arguments{
\item{dataOCC}{The species occurrence data. May be either a \code{SpatialPointsDataFrame}, \code{sf} or \code{data.frame} object.}

\item{type}{The type of observation model for the data. May be either: \code{PO}, \code{PA} or \code{Counts}.}

\item{varsOld}{The names of the old variables in the model which need to be converted to something new.}

\item{varsNew}{The name of the new variables in the model which will be used in the full model.}

\item{projection}{The CRS object to add to the species occurrence data.}

\item{boundary}{An sf object of the boundary of the study area, used to check if the data points are over the region.}
}
\value{
An \code{sf} object containing the locations of the species.
}
\description{
Function used to format structure data into a coherent framework.
}
