% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPriors.R
\name{getPriors}
\alias{getPriors}
\title{\code{getPriors}: Function to obtain priors from a fitted \code{\link[INLA]{inla}} object.}
\usage{
getPriors(model)
}
\arguments{
\item{model}{A \code{\link[INLA]{inla}} model object.}
}
\value{
The function returns a list with two items: fixedEffects describing the priors used for the fixed effects, and randomEffects describing the priors used for the random effects.
}
\description{
This function is used to obtain the prior distributions relevant to models fit with intSDM, and return them in a readable format.
}
