% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initValues.R
\name{initValues}
\alias{initValues}
\title{\code{initValues}: Function to obtain initial values from a \code{\link[PointedSDMs]{intModel}} object.}
\usage{
initValues(data, formulaComponents)
}
\arguments{
\item{data}{A \code{\link[PointedSDMs]{intModel}} object.}

\item{formulaComponents}{A vector of fixed effects for which to find initial values.}
}
\value{
The return of the function depends on the argument \code{Save} from the \code{startWorkflow} function. If this argument is \code{FALSE} then the objects will be saved to the specified directory. If this argument is \code{TRUE} then a list of different outcomes from the workflow will be returned.
}
\description{
This function is used to obtain initial values by running a separate linear model on each sub-likelihood of the model to find maximum likelihood estimates, and averaging the estimates across all likelihoods.
}
