% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainArea.R
\name{obtainArea}
\alias{obtainArea}
\title{obtainArea: Function to obtain a boundry for a specified country.}
\usage{
obtainArea(names, projection, ...)
}
\arguments{
\item{names}{A vector of names of countries used in the analysis.}

\item{projection}{Coordinate reference system used.}

\item{...}{Additional arguments passed to \link[giscoR]{gisco_get_countries}.}
}
\value{
An \code{sf} object of the boundary of the specified country.
}
\description{
Function to obtain a \code{sf} boundary object around a specified country.
}
