% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.PA_redlist.rda.R
\docType{data}
\name{PA_redlist}
\alias{PA_redlist}
\title{A dataset containing the presences and absences of three species: "Fraxinus excelsior", "Ulmus glabra" and "Arnica montana"}
\description{
A dataset containing the presences and absences of three species: "Fraxinus excelsior", "Ulmus glabra" and "Arnica montana"
}
\references{
\url{https://www.gbif.org}
}
\keyword{data}
