% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_structuredData.R
\name{show,structuredData-method}
\alias{show,structuredData-method}
\title{Print methods for structuredData object}
\usage{
\S4method{show}{structuredData}(object)
}
\arguments{
\item{object}{A "structedData" object.}
}
\description{
Print methods for structuredData object
}
