% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_structuredData.R
\docType{class}
\name{structuredData-class}
\alias{structuredData-class}
\title{bru_sdm_data class}
\value{
An S4 object with three slots, which contain lists of the data for the three observation models allowed in the modelling framework.
}
\description{
bru_sdm_data class
}
\section{Slots}{

\describe{
\item{\code{dataPO}}{A list of the present only datasets.}

\item{\code{dataPA}}{A list of the present absence datasets.}

\item{\code{dataCount}}{A list of the count datasets.}
}}

