% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structuredData.R
\name{structuredData}
\alias{structuredData}
\title{Function to convert a collection of datasets into an object which can be used in \code{speciesModel}.}
\usage{
structuredData(
  ...,
  datasetType = NULL,
  responsePA = NULL,
  trialsPA = NULL,
  speciesName = "species",
  coordinateNames = c("latitude", "longitude")
)
}
\arguments{
\item{...}{The datasets used in the model. May be either datasets or a SpatialPointsDataFrames.}

\item{datasetType}{A vector which gives the type of dataset. Must be either \code{'PO'} or \code{'PA'}. Defaults to \code{NULL}.}

\item{responsePA}{Name of the response variable in the \code{PA} datasets. Defaults to \code{NULL}.}

\item{trialsPA}{Name of the trial name variable in the \code{PA} datasets. Defaults to \code{NULL}.}

\item{speciesName}{Name of the species variable name. Defaults to \code{'species'}.}

\item{coordinateNames}{Name of the coordinates used in the model. Defaults to \code{c("latitude","longitude")}.}
}
\description{
Function to convert a collection of datasets into an object which can be used in \code{speciesModel}.
}
