% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmWorkflow.R
\name{sdmWorkflow}
\alias{sdmWorkflow}
\title{\code{sdmWorkflow}: Function to compile the reproducible workflow.}
\usage{
sdmWorkflow(Workflow = NULL)
}
\arguments{
\item{Workflow}{The \code{R6} object created from \code{startWorkflow}. This object should contain all the data and model information required to estimate and specify the model.}
}
\value{
The return of the function depends on the argument \code{Save} from the \code{startWorkflow} function. If this argument is \code{FALSE} then the objects will be saved to the specidfied directory. If this argument is \code{TRUE} then a list of different outcomes from the workflow will be returned.
}
\description{
This function is used to compile the reproducible workflow from the \code{R6} object created with \code{startFunction}. Depending on what was specified before, this function will estimate the integrated species distribution model, perform cross-validation, create predictions from the model and plot these predictions.
}
