% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_model.R
\name{species_model}
\alias{species_model}
\title{Function to run an SDM on select species, location and environmental covariates.}
\usage{
species_model(
  speciesNames,
  structuredData = NULL,
  spatialCovariates = NULL,
  worldclimCovariates = NULL,
  scale = FALSE,
  location = "Norway",
  boundary = NULL,
  return = "predictions map",
  mesh = NULL,
  meshParameters = NULL,
  spdeModel = NULL,
  projection = CRS("+proj=utm +zone=32 +ellps=WGS84 +datum=WGS84 +units=m +no_defs"),
  limit = 10000,
  options = list()
)
}
\arguments{
\item{speciesNames}{A vector of species' names to collect from GBIF.}

\item{structuredData}{Additional datasets to integrate with the presence only GBIF data. See the \code{structured_data} function. Defaults to \code{NULL}.}

\item{spatialCovariates}{Spatial covariates to include in the model. May be a \code{Raster} or \code{Spatial} object. Cannot be non-\code{NULL} if \code{worldclimCovariates} is non-\code{NULL}.}

\item{worldclimCovariates}{Names of the covariates to extract from Worldclim. Defaults to \code{NULL}; cannot be non-\code{NULL} if \code{spatialCovariates} is non-\code{NULL}.}

\item{scale}{Should the spatial covariates be scaled. Defaults to \code{FALSE}.}

\item{location}{Which area of Norway to model. Defaults to \code{'Norway'} which suggests a model for the entire county.}

\item{boundary}{SpatialPolygons object of the study area. If \code{NULL} an object may be formed with \code{location}.}

\item{return}{Object to return. Has to be one of \code{c('boundary' ,'species plot', 'mesh', 'mesh plot', 'model', 'predictions', 'predictions map')}.}

\item{mesh}{An inla.mesh object to include in the model. Defaults to \code{NULL}.}

\item{meshParameters}{A list of inla.mesh arguments to create a mesh if \code{mesh = NULL}.}

\item{spdeModel}{inla.spde model used in the model. May be a named list where the name of the spde object is the name of the associated dataset. Default NULL uses inla.spde2.matern.}

\item{projection}{CRS projection to use. Defaults to \code{CRS('+proj=utm +zone=32 +ellps=WGS84 +datum=WGS84 +units=m +no_defs')}.}

\item{limit}{Set the number of species downloaded. Defaults to \code{10000}.}

\item{options}{A list of \code{INLA} and \code{inlabru} options. Defaults to \code{NULL}.}
}
\description{
Function to run an SDM on select species, location and environmental covariates.
}
